Add option to use edge code when updating

+++ function/update.sh
@@ -5,8 +5,26 @@
 #
 
 demyx_update() {
-    demyx_echo 'Updating chroot.sh'
-    demyx_execute docker run -t --user=root --privileged --rm -v /usr/local/bin:/usr/local/bin demyx/utilities demyx-chroot
+    # Use latest code updates from git repo
+    if [[ "$DEMYX_BRANCH" = edge ]]; then
+        git clone https://github.com/demyxco/demyx.git /etc/demyx-edge
+        rm -rf /etc/demyx
+        mv /etc/demyx-edge /etc/demyx
+        chmod +x /etc/demyx/demyx.sh
+        chmod +x /etc/demyx/bin/demyx-api.sh
+        chmod +x /etc/demyx/bin/demyx-crond.sh
+        chmod +x /etc/demyx/bin/demyx-dev.sh
+        chmod +x /etc/demyx/bin/demyx-init.sh
+        chmod +x /etc/demyx/bin/demyx-prod.sh
+        chmod +x /etc/demyx/bin/demyx-skel.sh
+        chmod +x /etc/demyx/bin/demyx-ssh.sh
+    fi
+
+    # Don't update chroot.sh when hostname is code - for internal use only
+    if [[ "$DEMYX_HOST" != code ]]; then
+        demyx_echo 'Updating chroot.sh'
+        demyx_execute docker run -t --user=root --privileged --rm -v /usr/local/bin:/usr/local/bin demyx/utilities demyx-chroot
+    fi
 
     # Refresh stack if .env exists
     if [[ -f "$DEMYX_STACK"/.env ]]; then
